+++ source/Enum/Dhcpv6TargetScopeType.ps1
@@ -1,0 +1,6 @@
+enum Dhcpv6TargetScopeType
+{
+    Server = 1
+    Scope
+    Reservation
+}