+++ source/Enum/DynamicUpdatesType.ps1
@@ -1,0 +1,6 @@
+enum DynamicUpdatesType
+{
+    Always = 1
+    Never
+    OnClientRequest
+}