+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Unit tests for `DhcpServerDsc.OptionValueHelper`
 - Unit tests for `DSC_xDhcpServerReservation`
+- Resource `DhcpServerv4DnsDynamicUpdates`. Fixes [#65](https://github.com/dsccommunity/DhcpServerDsc/issues/65)
 
 ## [4.0.0] - 2025-01-20